refactor: restructure performance test scripts
