--- performance_test/clear_log.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-sudo chmod -R 777 logs
-
-sudo rm -rf logs/*
